requirements.txt
python-dotenv==1.0.1

numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
sentence-transformers==3.1.1
matplotlib==3.9.2
torch==2.4.1

langchain==0.3.1
openai==1.50.2 ## From the notebooks

pinecone-client==3.2.2
pinecone-text==0.9.0

tiktoken==0.7.0
tqdm==4.66.5

datasets==3.0.1
deepface==0.0.93